% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_fa2sf}
\alias{convert_fa2sf}
\title{Converts different formats to represent multiple sample data}
\usage{
convert_fa2sf(fa)
}
\arguments{
\item{fa}{the full abundance matrix with species in rows and samples in columns}
}
\value{
the sample frequency matrix
}
\description{
Converts the full abundance matrix into species frequencies
If S is the number of species and M is the number of samples, then fa is
the full abundance matrix of dimension S by M. The  for example
fa = [0 1 0;3 2 1;0 1 0] leads to sf = [0 1 0 2;3 2 1 1];
}
\references{
Haegeman, B. & R.S. Etienne (2017). A general sampling formula for community structure data. Methods in Ecology & Evolution. In press.
}
