% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fitresults}
\alias{fitresults}
\title{Maximum likelihood estimates and corresponding likelihood values for various fits to various tropical forest communities}
\format{A list of 20 lists, each containing either likelihood values or the corresponding parameter estimates. See description.}
\source{
Condit et al. (2002). Beta-diversity in tropical forest trees. Science 295: 666-669.
}
\usage{
data(fitresults)
}
\description{
Maximum likelihood estimates and corresponding likelihood values for various fits to various tropical forest communities, to test and illustrate the Independent Species approach.
\itemize{
\item fit1a.llikopt contains maximum likelihood values of fit of pm-dl model to dset1.abunvec
\item fit1a.parsopt contains maximum likelihood parameter estimates of fit of pm-dl model to dset1.abunvec
\item fit1b.llikopt contains maximum likelihood values of fit of pmc-dl model to dset1.abunvec
\item fit1b.parsopt contains maximum likelihood parameter estimates of fit of pmc-dl model to dset1.abunvec
\item fit2.llikopt contains maximum likelihood values of fit of rf-dl model to dset1.abunvec
\item fit2.parsopt contains maximum likelihood parameter estimates of fit of rf-dl model to dset1.abunvec
\item fit3.llikopt contains maximum likelihood values of fit of dd-dl model to dset1.abunvec
\item fit3.parsopt contains maximum likelihood parameter estimates of fit of dd-dl model to dset1.abunvec
\item fit4.llikopt contains maximum likelihood values of fit of pm-dl model to dset2.abunvec (multiple samples)
\item fit4.parsopt contains maximum likelihood parameter estimates of fit of pm-dl model to dset1.abunvec (multiple samples)
\item fit5.llikopt contains maximum likelihood values of fit of pm-dl model to dset3.abunvec (multiple guilds)
\item fit5.parsopt contains maximum likelihood parameter estimates of fit of pm-dl model to dset3.abunvec (multiple guilds)
\item fit6.llikopt contains maximum likelihood values of fit of pr-dl model to dset1.abunvec
\item fit6.parsopt contains maximum likelihood parameter estimates of fit of pr-dl model to dset1.abunvec
\item fit7.llikopt contains maximum likelihood values of fit of pm-dd model to dset1.abunvec
\item fit7.parsopt contains maximum likelihood parameter estimates of fit of pm-dd model to dset1.abunvec
\item fit8a.llikopt contains maximum likelihood values of fit of pm-dd model to dset4a.abunvec
\item fit8a.parsopt contains maximum likelihood parameter estimates of fit of pm-dd model to dset4a.abunvec
\item fit8b.llikopt contains maximum likelihood values of fit of pm-dd model to dset4b.abunvec
\item fit8b.parsopt contains maximum likelihood parameter estimates of fit of pm-dd model to dset4b.abunvec
}
}
\author{
Rampal S. Etienne & Bart Haegeman
}
\keyword{datasets}
