% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SADISA_test.R
\name{SADISA_test}
\alias{SADISA_test}
\title{Tests SADISA for data sets included in the paper by Haegeman & Etienne}
\usage{
SADISA_test(tol = 1e-06)
}
\description{
Tests SADISA for data sets included in the paper by Haegeman & Etienne
}
\references{
Haegeman, B. & R.S. Etienne (2017). A general sampling formula for community structure data. Methods in Ecology & Evolution. In press.
}
\keyword{model}
\keyword{species-abundance-distribution}
