% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SADISA_test.R
\name{SADISA_test}
\alias{SADISA_test}
\title{Tests SADISA for data sets included in the paper by Haegeman & Etienne}
\usage{
SADISA_test()
}
\description{
Tests SADISA for data sets included in the paper by Haegeman & Etienne
}
\references{
Haegeman, B. & Etienne, R.S. (2016). A general sampling formula for community abundance data. Methods in Ecology & Evolution. In review.
}
\keyword{model}
\keyword{species-abundance-distribution}

